Puts .5

+++ 0x05-pointers_arrays_strings/7-puts_half.c
@@ -34,7 +34,7 @@ int _strlen(char *s)
 
 void puts_half(char *str)
 {
-	int i = 0;
+	float i = 0;
 	float l;
 
 	l = _strlen(str) / 2;
